+++ apstools/plans/tests/test_alignment.py
@@ -43,6 +43,12 @@ for obj in (axis, m1, m2, noisy, scaler1, swait, calcs_enable):
 
 scaler1.select_channels()
 
+
+# TODO: Refactor 'pvoigt' as Python-only noisy pseudo-voigt signal,
+# a la 'ophyd.sim.noisy_det' to speed up testing of alignment plans.  
+# Use 'ophyd.sim.motor' as its positioner.
+# Currently, the 14 tests of 'lineup2()' take ~1m 40s.
+
 # First, must be connected to m1.
 pvoigt = SynPseudoVoigt(name="pvoigt", motor=axis, motor_field=axis.name)
 pvoigt.kind = "hinted"
@@ -342,14 +348,14 @@ _TestParameters = collections.namedtuple(
     "TestParameters", "peak base noise center sigma xlo xhi npts nscans tol outcome"
 )
 parms_signal_but_high_background = _TestParameters(1e5, 1e6, 10, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, False)
-parms_model_peak = _TestParameters(1e5, 0, 10, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, True)
+parms_model_peak = _TestParameters(1e5, 0, 10, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.06, True) #
 parms_high_background_poor_resolution = _TestParameters(1e5, 1e4, 10, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.1, True)
-parms_not_much_better = _TestParameters(1e5, 1e4, 10, 0.1, 0.2, -0.7, 0.5, 11, 2, 0.05, True)
+parms_not_much_better = _TestParameters(1e5, 1e4, 10, 0.1, 0.2, -0.7, 0.5, 11, 2, 0.08, True)
 parms_neg_peak_1x_base = _TestParameters(-1e5, -1e4, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.1, True)
-parms_neg_base = _TestParameters(1e5, -10, 10, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, True)
-parms_small_signal_zero_base = _TestParameters(1e-5, 0, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, True)
-parms_neg_small_signal_zero_base = _TestParameters(-1e5, 0, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, True)
-parms_small_signal_finite_base = _TestParameters(1e-5, 1e-7, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, True)
+parms_neg_base = _TestParameters(1e5, -10, 10, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.06, True) #
+parms_small_signal_zero_base = _TestParameters(1e-5, 0, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.06, True) #
+parms_neg_small_signal_zero_base = _TestParameters(-1e5, 0, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.06, True)
+parms_small_signal_finite_base = _TestParameters(1e-5, 1e-7, 1e-8, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.06, True) #
 parms_no_signal_only_noise = _TestParameters(0, 0, 1e-8, 0.1, 0.2, -1.0, 0.5, 11, 1, 0.05, False)
 parms_bkg_plus_noise = _TestParameters(0, 1, 0.1, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, False)
 parms_bkg_plus_big_noise = _TestParameters(0, 1, 100, 0.1, 0.2, -0.7, 0.5, 11, 1, 0.05, False)
@@ -363,11 +369,11 @@ parms_bkg_only__ZeroDivisionError = _TestParameters(0, 1, 0, 0.1, 0.2, -0.7, 0.5
         parms_signal_but_high_background,
         parms_model_peak,
         parms_high_background_poor_resolution,
-        parms_not_much_better,
+        parms_not_much_better,  # TODO: sigma
         parms_neg_peak_1x_base,
         parms_neg_base,
         parms_small_signal_zero_base,
-        parms_neg_small_signal_zero_base,
+        parms_neg_small_signal_zero_base,  # TODO: sigma
         parms_small_signal_finite_base,
         parms_no_signal_only_noise,
         parms_bkg_plus_noise,
@@ -404,7 +410,7 @@ def test_lineup2_signal_permutations(parms: _TestParameters):
     change_noisy_parameters()
 
     try:
-        centroid = stats._registers[detector.name].centroid
+        centroid = stats.analysis.centroid
         if parms.outcome:
             assert math.isclose(m1.position, centroid, abs_tol=parms.tol)
             assert math.isclose(parms.center, centroid, abs_tol=parms.tol)